src/endpoints/vercel_gemini_chatbot/requirements.txt
fastapi
uvicorn
google-generativeai
requests
python-dotenv